requirements.txt
# Core cryptography
cryptography>=42.0.0

# Image handling (for image transfer feature)
Pillow>=10.0.0

# System tray integration (optional)
pystray>=0.19.0

# CLI interface (for colored terminal output)
colorama>=0.4.6

# Testing and code quality
coverage>=7.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code quality tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pylint>=3.0.0
isort>=5.12.0

# Note: asyncio is built into Python 3.7+
# Note: tkinter is built into Python (for GUI client)
